+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.9
 - Enable DAP customizations by default and add them to the Settings UI.
   - Improve **Set Value** behavior in sevaral scenarios.
+- Rename `tracciatto.patchNilVariableExpansion` to `tracciatto.patchSimpleTypeExpansion` to accomodate for more types (check the README).
 - Fix: `rdbg` debug-type support can only be enabled if **vscode-rdbg** is not installed.
 - Fix: Blank icon for **Exception Filters**.